+++ server/main.js
@@ -119,7 +119,7 @@ const srv = http.createServer(async (req, res) => {
 
     getIMFFallbackResults({
         $p,
-        basePath: `${basePath}`,
+        basePath,
         langData: await getJSON(userParamsWithDefaults.languages),
         async resultsDisplay (resultsArgs, ...args) {
             const serverOutput = $p.get('serverOutput', true);
@@ -142,4 +142,8 @@ const srv = http.createServer(async (req, res) => {
         }
     });
 });
-srv.listen(port);
+if (!userParams.basePath) {
+    srv.listen(port);
+} else {
+    srv.listen(port, userParams.basePath);
+}